IFS: Move generic functions to sandstone_ifs.h
